sadak-saathi-backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
redis
python-dotenv
